+++ docs/changelog.md
@@ -6,7 +6,7 @@ order: 100
 
 # Changelog
 
-## 4.4.0
+## 5.0.0
 
 ### Added
 
@@ -17,7 +17,9 @@ order: 100
   - A registry means developers can now fetch all blocks registered for a certain type of use. Supports multiple contexts per block.
   - Implemented `GetBlocksForContext()` and `CanBlockBeUsedInContext()` API functions for context-aware block management
 - Documentation tests now cover mandatory YAML frontmatter for all docs. Files with no yaml or missing `title` or `order` fields will break the docs service.
+### Changed
 
+- **Breaking!** Removed `GetRegisteredBlocks` from blocks package in favour of context system.
 [Full Changelog](https://github.com/nathanhollows/Rapua/releases/tag/v4.4.0)
 
 ## 4.3.0 (2025-10-09)